Added logs, synced heartbest and interval

+++ src/main/java/com/dreamlab/InsertBlocks.java
@@ -31,7 +31,7 @@ public final class InsertBlocks {
         final String metadataDirectory = args[2];
         final File blocksDir = new File(blocksDirectory);
         final File metadataDir = new File(metadataDirectory);
-        final int interval = Integer.parseInt(args[3]);
+        final double interval = Double.parseDouble(args[3]);
         final List<String> blocks = Arrays.stream(blocksDir.listFiles()).map(File::getAbsolutePath).collect(Collectors.toList());
         final List<String> metadata = Arrays.stream(metadataDir.listFiles()).map(File::getAbsolutePath).collect(Collectors.toList());
         Collections.sort(blocks);
@@ -59,7 +59,7 @@ public final class InsertBlocks {
                 final long end = System.currentTimeMillis();
                 LOGGER.info(String.format("[Client] [Outer %s] EdgeServer.putBlockAndMetadata: %d", blockId, (end - start)));
                 LOGGER.info("Success: " + Utils.getUuidFromMessage(blockIdResponse.getBlockId()));
-                final long sleepTime = interval * 1000L - (end - start);
+                final long sleepTime = (long) (interval * 1000L - (end - start));
                 Thread.sleep(sleepTime >= 0? sleepTime : 0);
             }
             catch (Exception ex) {
