+++ src/main/java/com/dreamlab/server/EdgeServer.java
@@ -17,7 +17,7 @@ public class EdgeServer {
         final String edgeIp = System.getenv("device_ip");
         System.out.println("Edge Server Starting on " + edgeIp);
         final int edgePort = Integer.parseInt(args[0]);
-        final int heartbeatTtlSecs = Integer.parseInt(args[1]);
+        final double heartbeatTtlSecs = Double.parseDouble(args[1]);
         final String fogsConfigFilePath = args[2];
         final String trajectoryFilePath = args[3];
         Server server = ServerBuilder
